Feat(dx): migrate scripts to src/, integrate Vite bundling, add coverage
Move all client-side scripts from public/scripts/ to src/scripts/ so they
go through Vite's transform pipeline — enabling bundling, minification,
cache-busting hashes, and Istanbul instrumentation.

- Remove IIFE wrappers (ES module scope provides the same isolation)
- Update Astro components to use native <script> imports; Vite now
  bundles, deduplicates, and processes them (7 → 15 modules at build)
- Inline nojs.js one-liner directly in head.astro
- Add @cypress/code-coverage + vite-plugin-istanbul + nyc
- Gate instrumentation behind CYPRESS_COVERAGE=true (no dev-server overhead)
- Add test:coverage script; baseline coverage: 68.5% stmts / 48% branch
