+++ astro.config.mjs
@@ -3,6 +3,7 @@ import preact from '@astrojs/preact';
 import tailwindcss from '@tailwindcss/vite';
 import { defineConfig } from 'astro/config';
 import { imagetools } from 'vite-imagetools';
+import istanbul from 'vite-plugin-istanbul';
 
 // https://astro.build/config
 export default defineConfig({
@@ -15,6 +16,17 @@ export default defineConfig({
 
   vite: {
     // @ts-ignore - Plugin compatibility issues with newer Vite versions
-    plugins: [tailwindcss(), imagetools()],
+    plugins: [
+      tailwindcss(),
+      imagetools(),
+      istanbul({
+        include: 'src/**/*',
+        exclude: ['node_modules', 'cypress'],
+        extension: ['.js', '.ts', '.jsx', '.tsx', '.astro'],
+        cypress: true,
+        requireEnv: true,
+        forceBuildInstrument: false,
+      }),
+    ],
   },
 });
