+++ cypress.config.js
@@ -1,11 +1,12 @@
+import codeCoverage from '@cypress/code-coverage/task';
 import { defineConfig } from 'cypress';
 
 export default defineConfig({
   allowCypressEnv: false,
   e2e: {
-    // eslint-disable-next-line no-unused-vars
     setupNodeEvents(on, config) {
-      // implement node event listeners here
+      codeCoverage(on, config);
+      return config;
     },
   },
 });
