+++ package.json
@@ -14,6 +14,7 @@
     "lint:prettier:write": "prettier --write .",
     "cy:run": "cypress run --config baseUrl=http://localhost:4321",
     "test:e2e": "start-server-and-test dev http://localhost:4321 cy:run",
+    "test:coverage": "CYPRESS_COVERAGE=true start-server-and-test dev http://localhost:4321 cy:run",
     "test": "npm run lint && npm run test:e2e",
     "prepare": "npx husky",
     "security:audit": "npm audit",
@@ -27,6 +28,7 @@
   },
   "devDependencies": {
     "@astrojs/preact": "^4.1.3",
+    "@cypress/code-coverage": "^4.0.3",
     "@tailwindcss/postcss": "^4.2.2",
     "@tailwindcss/vite": "^4.2.2",
     "@typescript-eslint/eslint-plugin": "^8.56.1",
@@ -39,6 +41,7 @@
     "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-react": "^7.37.5",
     "husky": "^9.1.7",
+    "nyc": "^18.0.0",
     "patch-package": "^8.0.1",
     "postcss": "^8.5.6",
     "preact": "^10.29.1",
@@ -50,6 +53,7 @@
     "typescript": "^5.9.3",
     "typescript-eslint": "^8.56.1",
     "vite": "^8.0.7",
-    "vite-imagetools": "^10.0.0"
+    "vite-imagetools": "^10.0.0",
+    "vite-plugin-istanbul": "^8.0.0"
   }
 }
