--- public/scripts/nojs.js
@@ -1,1 +1,0 @@
-document.querySelector('html')?.classList.remove('nojs');
