+++ src/components/cv/cv.astro
@@ -42,4 +42,6 @@ const { id, entries = [] }: Props = Astro.props as Props;
   </ol>
 </section>
 
-<script type="module" src="/scripts/animate-sections.js"></script>
+<script>
+  import '../../scripts/animate-sections.js';
+</script>
