+++ src/components/header.astro
@@ -68,5 +68,7 @@ const { href }: Props = Astro.props;
   </div>
 </header>
 
-<script type="module" src="/scripts/particles.js"></script>
-<script type="module" src="/scripts/tilt.js"></script>
+<script>
+  import '../scripts/particles.js';
+  import '../scripts/tilt.js';
+</script>
