+++ src/components/testimonials/testimonials.astro
@@ -104,4 +104,6 @@ const testimonialCount = testimonials.length;
   </div>
 </section>
 
-<script type="module" src="/scripts/testimonials.js"></script>
+<script>
+  import '../../scripts/testimonials.js';
+</script>
